requirements.txt
absl-py==2.0.0
accelerate==0.24.1
alembic==1.12.0
astunparse==1.6.3
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
diffusers==0.21.4
elasticsearch==7.17.3
filelock==3.12.4
Flask==3.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.9.2
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Mako==1.2.4
Markdown==3.5
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
Pillow==10.0.1
protobuf==4.24.4
psutil==5.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
six==1.16.0
SQLAlchemy==2.0.22
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.34.0
typing_extensions==4.8.0
urllib3==1.26.17
Werkzeug==2.3.0
wrapt==1.14.1
zipp==3.17.0
